component {reactR}	R Documentation

Create a React component

Description

Create a React component

Usage

component(name, varArgs = list())

Arguments

name
Name of the React component, which must start with an upper-case character.

varArgs
Attributes and children of the element to pass along to tag as varArgs.

Value

An htmltools tag object

Examples

component("ParentComponent",
  list(
    x = 1,
    y = 2,
    component("ChildComponent"),
    component("OtherChildComponent")
  )
)

[Package reactR version 0.6.1 Index]