reactMarkup {reactR}	R Documentation

Prepare data that represents a single-element character vector, a React component, or an htmltools tag for sending to the client.

Description

Tag lists as returned by htmltools tagList are not currently supported.

Usage

reactMarkup(tag)

Arguments

tag
character vector or React component or tag

Value

A reactR markup object suitable for being passed to createWidget as widget instance data.

[Package reactR version 0.6.1 Index]