scaffoldReactShinyInput {reactR}	R Documentation

Create implementation scaffolding for a React.js-based Shiny input.

Description

Add the minimal code required to implement a React.js-based Shiny input to an R package.

Usage

scaffoldReactShinyInput(name, npmPkgs = NULL, edit = interactive())

Arguments

name
Name of input

npmPkgs
Optional NPM packages upon which this input is based which will be used to populate package.json. Should be a named list of names to versions.

edit
Automatically open the input's source files after creating the scaffolding.

Note

This function must be executed from the root directory of the package you wish to add the input to.

[Package reactR version 0.6.1 Index]