scaffoldReactWidget {reactR}	R Documentation

Create implementation scaffolding for a React.js-based HTML widget

Description

Add the minimal code required to implement a React.js-based HTML widget to an R package.

Usage

scaffoldReactWidget(name, npmPkgs = NULL, edit = interactive())

Arguments

name
Name of widget

npmPkgs
Optional NPM packages upon which this widget is based which will be used to populate package.json. Should be a named list of names to versions.

edit
Automatically open the widget's JavaScript source file after creating the scaffolding.

Note

This function must be executed from the root directory of the package you wish to add the widget to.

[Package reactR version 0.6.1 Index]